Turtle crossing
Frogger type game where you play as a turtle trying to cross a street as cars race by. difficulty scales with level, increasing the amount and speed of the cars. there are slow, medium, and fast cars, and an even distribution of each. the game is playable, but experiences tremendous slowdown. there is certainly room for optimization